refactor: 카테고리 관련 테이블 구조 변경으로 인한 수정

+++ src/main/java/com/application/poppool/domain/sign_up/dto/request/SignUpRequest.java
@@ -15,6 +15,6 @@ public class SignUpRequest {
     private int age;
     private String socialEmail;
     private SocialType socialType;
-    private Set<Long> interestCategories;
+    private Set<Integer> interestCategories;
 
 }
